requirements.txt
aiogram
telethon
aiosqlite
aiohttp
python-dotenv
cryptg
Pillow
moviepy
apscheduler